refactor: 플레이어 및 딜러간 승부 계산 기능 Gamer 에서 분리
