+++ src/main/java/domain/Gamer.java
@@ -23,18 +23,4 @@ public class Gamer {
     public SummationCardPoint getSummationCardPoint() {
         return holdingCards.calculateTotalPoint();
     }
-
-    public GameResult getGameResult(Gamer other) {
-        if (this.getSummationCardPoint().isBiggerThan(other.getSummationCardPoint())) {
-            return GameResult.WIN;
-        }
-        if (this.getSummationCardPoint().equals(other.getSummationCardPoint())) {
-            return GameResult.TIE;
-        }
-        return GameResult.LOSE;
-    }
-
-    public String getName() {
-        return name;
-    }
 }
